Add machinery for storing plays
This involves some updates to the track template that will populate and
submit the next form for me with interesting data about what the user
has supplied by way of feedback, like the amount of time they listened.
Then any time the user requests a new track we store that they played
the track. This means we aren't really exercising the 'time when
reinforced' yet, but we will by storing it client side. Eventually.

I've also moved to storing and retrieving the mood from the database so
that as I modify code I can keep the mood around