Massive update to use storage
Now I've got storage tables. I want therefore to start actually treating
the track location and the track as two separate things. This may seem
wasteful, but it will make a big difference to the final architecture of
the program. For now I have some nice high-level information on the
track by using eyed3 to get the tags out of the music files.

I also now have a list of all of the tracks tied to the storage